requirements.txt
backports.entry-points-selectable==1.1.1
cfgv==3.3.1
click==8.0.3
dicttoxml>=1.7.4
distlib==0.3.4
filelock==3.4.0
Flask==2.0.2
flask-restful>=0.3.9
identify==2.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
nodeenv==1.6.0
platformdirs==2.4.0
pre-commit==2.16.0
PyYAML==6.0
six==1.16.0
toml==0.10.2
virtualenv==20.10.0
Werkzeug==2.0.2
gevent==21.12.0
mysql-connector-python==8.0.33

